move all vote buttons into PostActionBar
